[refactor] [#2] 우측 상단 버튼들 스타일 수정
- 우측 상단 버튼들의 스타일을 불러오는 경로 수정
- 우측 상단 버튼 컨테이너 폴더 깊이 한 단계 낮춤
- README.md파일 업데이트
