+++ README.md
@@ -1,4 +1,9 @@
-# Project 08
+<p align="center">
+    <img width="549" alt="스크린샷 2020-11-19 오후 10 42 21" src="https://user-images.githubusercontent.com/26402298/99673762-8a71e080-2ab8-11eb-84ca-909703fae826.png">
+</p>
+<p align="center">
+    <img src="https://github.com/qkrdmstlr3/svg-icon-animation/blob/master/map-icon/map-icon.svg" height="300" />
+</p>
 
 ### 폴더 구조
