+++ src/components/Map/Button/ButtonPresenter.tsx
@@ -1,5 +1,5 @@
 import React, { ReactElement, MouseEvent } from 'react';
-import styled from '@emotion/styled';
+import styled from '../../../utils/styles/styled';
 
 interface ButtonPropsInterface {
   width?: string;
